+++ test/Agent.test.js
@@ -1,8 +1,23 @@
 var assert = require('assert');
+var Promise = require('promise');
 var Agent = require('../lib/Agent');
 var system = require('../lib/system');
 var LocalTransport = require('../lib/transport/local/LocalTransport');
 var DistribusTransport = require('../lib/transport/distribus/DistribusTransport');
+var WebSocketTransport = require('../lib/transport/websocket/WebSocketTransport');
+
+/**
+ * Get a free local port
+ * @returns {Promise.<number>} Resolves with a free port number
+ */
+function freeport () {
+  return new Promise(function (resolve, reject) {
+    var f = require('freeport');
+    f(function (err, port) {
+      err ? reject(err) : resolve(port);
+    })
+  });
+}
 
 describe('Agent', function() {
 
@@ -242,6 +257,53 @@ describe('Agent', function() {
 
       // send messages to agents connected via a different transport
       agent2.send('agent3', 'hello'); // should go over local1
+    });
+
+    it('should send a message via websocket transport', function () {
+      var transport1;
+      var transport2;
+      var agent1;
+      var agent2;
+      var conn1;
+      var conn2;
+
+      return Promise.all([freeport(), freeport()])
+          .then(function (ports) {
+            transport1 = new WebSocketTransport({url: 'ws://localhost:' + ports[0] + '/agents/:id'});
+            transport2 = new WebSocketTransport({url: 'ws://localhost:' + ports[1] + '/agents/:id'});
+
+            return Promise.all([transport1.ready, transport2.ready]);
+          })
+          .then(function () {
+            agent1 = new Agent('agent1');
+            agent2 = new Agent('agent2');
+
+            conn1 = agent1.connect(transport1);
+            conn2 = agent2.connect(transport2);
+
+            return Promise.all([conn1.ready, conn2.ready]);
+          })
+          .then(function () {
+            return new Promise(function (resolve, reject) {
+              agent1.receive = function (from, message) {
+                try {
+                  assert.equal(from, conn2.url);
+                  assert.equal(message, 'hello');
+                  resolve();
+                }
+                catch (err) {
+                  reject(err);
+                }
+              };
+
+              // send messages to the agent on an other websocket server
+              agent2.send(conn1.url, 'hello');
+            });
+          })
+          .then(function () {
+            transport1.close();
+            transport2.close();
+          });
     });
 
   });